fix: remove bloom in 3d render
three@0.155.0 changed the color space where effects are applied. Bloom
applies in linear color space instead of screen sRGB space [1] (see [2]
for the tempting fix to put it back to work in  sRGB space), which made
it turn out very bloomy.

The bloom effect seemed to be mostly on reflection from the environment.
The emissive lights on the light bulbs are actually rather weak
comparatively, so their bloom isn't so perceptible.

Pre-155, the bloom effect wasn't really perceptible anyway, so I figured
I would remove all of it.

[1] https://github.com/mrdoob/three.js/wiki/Migration-Guide#154--155
[2] https://github.com/pmndrs/postprocessing/issues/370