fix: make the water toxic again
This change fixes a regression in water color from three 0.151
to 0.152. There is at least one more color regression to 0.155 not fixed
here.

three@0.152.0 changed the way color management is performed [1]. This
barely affects us because React Three Fiber already set up color
management in the renderer to use the new 0.152.0 defaults, but it did
affect the one mesh material we define in code, changing the color space
in which the hex the color string is interpreted.

One possible explanation is that color management got enabled late by
three-js before 0.152.0 made it enabled by default, relative to the
water material definition. This would explain why the color is not the
same after hot module reloading.

[1] https://discourse.threejs.org/t/updates-to-color-management-in-three-js-r152/50791/1