chore: bump next to 14.0.2
This fixes a bug in the compiled output of `next build`. The `prefetch`
prop of `next-intl/link` got changed from `undefined` to `false` by an
overzealous optimization.

That resulted in a forced reload of the page on every navigation through
a next-intl link. We use that for all links, so it effectively turned
our single-page app into a multi-page one :')