fix: use an optimized tiny environment map
No need to download 1.5 megs of high resolution photography, a few
pixels is enough to set the mood. We don't see the low resolution in
reflections.

The resampling causes a small difference in the render, but nothing
major.

```sh
convert san_giuseppe_bridge_1k.hdr -scale '64x' san_giuseppe_bridge_tiny.hdr
```

(anything under that size does not seem to be tolerated by three.js and
results in the environment not showing, and a very dark scene)