fix: replace zod parse with safeParse, and control error payload
This more verbose approach lets me use a custom error class whose
message is a string representation of the error object. This means the
Next server will output the full depth of the error payload.

That way, if one item is invalid in the schedule, I can tell which one:
instead of `path: [Array]` I will see the actual path of the payload
that was faulty.